fix(modes): improve Mode A/C decoding and rematching
- fix int32 overflow in ModeAToModeC causing negative Gillham altitudes to appear as large positives on 64-bit platforms.
- add an explicit int32 cast to match C behavior.
- add upstream-compatible Mode A/C rematching in PeriodicUpdate.
- match dump1090-mutability track.c re-evaluation behavior.
- promote the DataValidity source to SOURCE_MODE_S for Mode A/C squawk and altitude so dataValid returns true.
- add a --modeac flag to gate Mode A/C demodulation, disabled by default.
- add tests for Gillham altitude, squawk and altitude fields, SPI ident, synthetic ICAO address, periodic rematch, and the disabled-by-default config.

+++ modes/mode_ac.go
@@ -73,7 +73,9 @@ func ModeAToModeC(modeA uint32) int {
 		oneHundreds = 6 - oneHundreds
 	}
 
-	return int((fiveHundreds * 5) + oneHundreds - 13)
+	// Cast through int32 to match C behavior where implicit int truncation
+	// on 32-bit systems causes negative altitude wraparound.
+	return int(int32((fiveHundreds * 5) + oneHundreds - 13))
 }
 
 // DecodeID13Field decodes a 13-bit identity (squawk) field.
